+++ bika/lims/upgrade/v01_03_000.py
@@ -188,6 +188,10 @@ def upgrade(tool):
     # https://github.com/senaite/senaite.core/pull/1230
     update_ar_listing_catalog(portal)
 
+    # Updates Indexes/Metadata of the bika_catalog
+    # https://github.com/senaite/senaite.core/pull/1231
+    update_bika_catalog(portal)
+
     logger.info("{0} upgraded to version {1}".format(product, version))
     return True
 
@@ -1624,6 +1628,7 @@ def update_ar_listing_catalog(portal):
     """Add Indexes/Metadata to bika_catalog_analysisrequest_listing
     """
     cat_id = CATALOG_ANALYSIS_REQUEST_LISTING
+
     catalog = api.get_tool(cat_id)
 
     logger.info("Updating Indexes/Metadata of Catalog '{}'".format(cat_id))
@@ -1635,6 +1640,34 @@ def update_ar_listing_catalog(portal):
 
     metadata_to_add = [
         "getClientID",
+    ]
+
+    for index in indexes_to_add:
+        add_index(portal, cat_id, *index)
+
+    for metadata in metadata_to_add:
+        refresh = metadata not in catalog.schema()
+        add_metadata(portal, cat_id, metadata, refresh_catalog=refresh)
+
+
+def update_bika_catalog(portal):
+    """Add Indexes/Metadata to bika_catalog
+    """
+    cat_id = "bika_catalog"
+
+    catalog = api.get_tool(cat_id)
+
+    logger.info("Updating Indexes/Metadata of Catalog '{}'".format(cat_id))
+
+    indexes_to_add = [
+        # name, attribute, metatype
+        ("getClientID", "getClientID", "FieldIndex"),
+        ("getClientBatchID", "getClientBatchID", "FieldIndex"),
+    ]
+
+    metadata_to_add = [
+        "getClientID",
+        "getClientBatchID",
     ]
 
     for index in indexes_to_add: